Phase 4: ww_executor_run C ABI entry (4e)
Publish the executor through the flat C ABI so a Java + Panama consumer
(or any other FFI binder) can drive a run without touching the C++
types directly.

worldwalker_c.h gains the executor wire-shapes: WwTile / WwGoal /
WwEvent / WwCapabilityEntry / WwCapabilitySnapshot as PascalCase POD
structs, WW_STATUS_* and WW_EVENT_* int32 sentinels, ten function-
pointer typedefs covering reads / actions / control / progress, the
WwCallbacks vtable, and ww_executor_run(artifact, pool, goal, cb) ->
int32 returning one of the WW_STATUS_* values.

Callbacks.h is reduced to a C++ overlay: it includes the C header,
re-publishes the structs and function-pointer typedefs into ww::exec
via using-declarations, keeps enum class WwStatus / WwEventKind for
type-safe construction on the C++ side, and pins each enumerator to
its WW_* macro via static_assert so the two views never drift. The
struct sizes already carry static_asserts (12/16/16/8). Callbacks
itself becomes `using Callbacks = ::WwCallbacks;` so existing C++ call
sites compile unchanged. Executor.cpp adds one cast at emit() where
WwEventKind is brace-init into WwEvent::kind (int32 on the wire), and
the harness onEvent does one cast at the boundary before its enum
comparisons.

ww_executor_run null-checks the artifact, the pool, the vtable, and
each required vtable entry (everything except onEvent) so a missing
function pointer fails with a clean WW_STATUS_FAILED + last-error
instead of crashing inside the executor's tight loop. The Executor
itself is constructed in-place from the borrowed reader and pool,
run(goal) is invoked synchronously, and any std::exception is funneled
to setLastError + WW_STATUS_FAILED.

The harness gains a Test 5 in dumpExecutor that opens a second
artifact + pool through ww_artifact_open / ww_context_pool_create,
runs the same farthest-tile walk-only goal as Test 2 via
ww_executor_run, then closes both handles. Results match Test 2
byte-for-byte: status=0 walks=2 sleeps=2 cancels=2 reads=5 events=3
abort=0, landed at goal. Tests 1-4 are unchanged from the 4d
baseline. dumpArtifact / dumpExecutor now take the artifact path so
the FFI test can re-open it through the C surface.

+++ src/exec/Executor.cpp
@@ -85,7 +85,7 @@ namespace ww::exec
         {
             return;
         }
-        const WwEvent event{ kind, 0, stepIndex, transitionIndex };
+        const WwEvent event{ static_cast<int32_t>(kind), 0, stepIndex, transitionIndex };
         callbacks->onEvent(callbacks->user, &event);
     }
